server/pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodalrag"
version = "0.1.0"
description = "Django REST backend for multimodal RAG with Neo4j, MLX, and more"
authors = [{ name = "Your Name" }]
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.15",
    "django-cors-headers>=4.4",
    "pydantic>=2.8",
    "numpy>=1.26",
    "scipy>=1.11",
    "pillow>=10.4",
    "tqdm>=4.66",
    "opencv-python-headless>=4.10",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0",
    "PyMuPDF>=1.24",
    "regex>=2024.7.24",
    "python-magic>=0.4.27",
    "torch>=2.3",
    "neo4j~=5.28",
    "faiss-cpu>=1.8.0",
    "uvicorn>=0.30",
    "fastapi>=0.115",
    "mlx-lm>=0.27",
    "mlx-vlm>=0.0.11",
    "mlx>=0.29",
    "transformers>=4.44",
    "tokenizers>=0.19"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["core*"]